chore(planner): Move dominator tests out of planner module implementation
